requirements/base.txt
django-cms>=3.6
django-filer>=1.4
django-mptt>=0.10
django-parler>=1.9
